add header and member component
