+++ src/components/MemberComponents/MemberComponents.jsx
@@ -1,90 +1,20 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import SwipeableViews from 'react-swipeable-views';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-import {
-    Button,
-    Form,
-    FormGroup,
-    Label,
-    Input,
-    Container,
-    Row,
-    Col,
-} from 'reactstrap';
-import Order from '../Order/Order';
-import MyService from '../MyService/MyService';
-import PurchaseHistory from '../PurchaseHistory/PurchaseHistory';
-
-function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import MemberGuest from './MemberGuest';
+import MemberTalent from './MemberTalent';
+import { fetchGetMember } from '../../redux/actions';
+function MemberComponents(props) {
+    const dispatch = useDispatch();
+    const member = useSelector((state) => state.getmember.role);
+    console.log(member, 'di member component');
+    useEffect(() => {
+        dispatch(fetchGetMember());
+    }, []);
     return (
-        <div
-            role="tabpanel"
-            hidden={value !== index}
-            id={`full-width-tabpanel-${index}`}
-            aria-labelledby={`full-width-tab-${index}`}
-            {...other}
-        >
-            {value === index && (
-                <Box p={3}>
-                    <Typography>{children}</Typography>
-                </Box>
-            )}
-        </div>
+        <React.Fragment>
+            {member === 'TALENT' ? <MemberTalent /> : <MemberGuest />}
+        </React.Fragment>
     );
 }
 
-TabPanel.propTypes = {
-    children: PropTypes.node,
-    index: PropTypes.any.isRequired,
-    value: PropTypes.any.isRequired,
-};
-
-const useStyles = makeStyles((theme) => ({
-    root: {
-        backgroundColor: theme.palette.background.paper,
-    },
-}));
-
-export default function MemberComponent() {
-    const classes = useStyles();
-    const theme = useTheme();
-    const [value, setValue] = React.useState(0);
-
-    const handleChange = (event, newValue) => {
-        setValue(newValue);
-    };
-
-    const handleChangeIndex = (index) => {
-        setValue(index);
-    };
-
-    return (
-        <div className={classes.root}>
-            <Tabs
-                value={value}
-                onChange={handleChange}
-                indicatorColor="primary"
-                textColor="primary"
-                centered
-            >
-                <Tab label="1. Purchase History" />
-            </Tabs>
-            <SwipeableViews
-                axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
-                index={value}
-                onChangeIndex={handleChangeIndex}
-            >
-                <TabPanel value={value} index={0} dir={theme.direction}>
-                    <PurchaseHistory />
-                </TabPanel>
-            </SwipeableViews>
-        </div>
-    );
-}
+export default MemberComponents;
