+++ src/components/MemberComponents/MemberTalent.jsx
@@ -1,0 +1,98 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import SwipeableViews from 'react-swipeable-views';
+import { makeStyles, useTheme } from '@material-ui/core/styles';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
+import Typography from '@material-ui/core/Typography';
+import Box from '@material-ui/core/Box';
+import {
+    Button,
+    Form,
+    FormGroup,
+    Label,
+    Input,
+    Container,
+    Row,
+    Col,
+} from 'reactstrap';
+import Order from '../Order/Order';
+import MyService from '../MyService/MyService';
+import PurchaseHistory from '../PurchaseHistory/PurchaseHistory';
+
+function TabPanel(props) {
+    const { children, value, index, ...other } = props;
+
+    return (
+        <div
+            role="tabpanel"
+            hidden={value !== index}
+            id={`full-width-tabpanel-${index}`}
+            aria-labelledby={`full-width-tab-${index}`}
+            {...other}
+        >
+            {value === index && (
+                <Box p={3}>
+                    <Typography>{children}</Typography>
+                </Box>
+            )}
+        </div>
+    );
+}
+
+TabPanel.propTypes = {
+    children: PropTypes.node,
+    index: PropTypes.any.isRequired,
+    value: PropTypes.any.isRequired,
+};
+
+const useStyles = makeStyles((theme) => ({
+    root: {
+        backgroundColor: theme.palette.background.paper,
+    },
+}));
+
+export default function TalentComponent() {
+    const classes = useStyles();
+    const theme = useTheme();
+    const [value, setValue] = React.useState(0);
+
+    const handleChange = (event, newValue) => {
+        setValue(newValue);
+    };
+
+    const handleChangeIndex = (index) => {
+        setValue(index);
+    };
+
+    return (
+        <div className={classes.root}>
+            <Tabs
+                value={value}
+                onChange={handleChange}
+                indicatorColor="primary"
+                textColor="primary"
+                centered
+            >
+                <Tab label="1. Purchase History" />
+                <Tab label="2. My Service" />
+                <Tab label="3. Order" />
+            </Tabs>
+            <SwipeableViews
+                axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
+                index={value}
+                onChangeIndex={handleChangeIndex}
+            >
+                <TabPanel value={value} index={0} dir={theme.direction}>
+                    <PurchaseHistory />
+                </TabPanel>
+                <TabPanel value={value} index={1} dir={theme.direction}>
+                    <MyService />
+                </TabPanel>
+                <TabPanel value={value} index={2} dir={theme.direction}>
+                    <Order />
+                </TabPanel>
+            </SwipeableViews>
+        </div>
+    );
+}
